figure fixed for chart page, turned out the library input of bohek on tutorial was outdated.
