+++ src/templates/base.html
@@ -8,6 +8,17 @@
   <link rel="stylesheet" href="{{ url_for("static", filename="css/skeleton.css") }}">
   <link rel="stylesheet" href="{{ url_for("static", filename="css/style.css") }}">
   <link rel="icon" type="image/png" href="{{ url_for("static", filename="assets/dino.png") }}">
+  <link
+      href="https://cdn.pydata.org/bokeh/release/bokeh-1.0.2.min.css"
+      rel="stylesheet" type="text/css">
+  <link
+      href="https://cdn.pydata.org/bokeh/release/bokeh-widgets-1.0.2.min.css"
+      rel="stylesheet" type="text/css">
+  <link
+      href="https://cdn.pydata.org/bokeh/release/bokeh-tables-1.0.2.min.css"
+      rel="stylesheet" type="text/css">
+
+
   <title>Flask Demo</title>
 </head>
 <body>
