+++ src/templates/chart.html
@@ -4,15 +4,23 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <link href="http://cdn.pydata.org/bokeh/release/bokeh-0.12.5.min.css" rel="stylesheet">
-  <link href="http://cdn.pydata.org/bokeh/release/bokeh-widgets-0.12.0.min.css" rel="stylesheet">
+  <link
+      href="https://cdn.pydata.org/bokeh/release/bokeh-1.0.2.min.css"
+      rel="stylesheet" type="text/css">
+  <link
+      href="https://cdn.pydata.org/bokeh/release/bokeh-widgets-1.0.2.min.css"
+      rel="stylesheet" type="text/css">
+  <link
+      href="https://cdn.pydata.org/bokeh/release/bokeh-tables-1.0.2.min.css"
+      rel="stylesheet" type="text/css">
   <title>Bar charts with Bokeh!</title>
 </head>
 <body>
-  <h1>market performance plot of {{ stock_name}} </h1>
+  <h1>market performance plot of {{ stock_name }} </h1>
   {{ the_div|safe }}
-  <script src="http://cdn.pydata.org/bokeh/release/bokeh-0.12.5.min.js"></script>
-  <script src="http://cdn.pydata.org/bokeh/release/bokeh-widgets-0.12.5.min.js"></script>
+  <script src="http://cdn.pydata.org/bokeh/release/bokeh-1.0.2.min.js"></script>
+  <script src="http://cdn.pydata.org/bokeh/release/bokeh-widgets-1.0.2.min.js"></script>
+  <script src="https://cdn.pydata.org/bokeh/release/bokeh-tables-1.0.2.min.js"></script>
   {{ the_script|safe }}
 </body>
 </html>
