+++ src/templates/portfolio/preview.html
@@ -14,5 +14,7 @@
     {{ form.portfolios.label }}{{ form.portfolios() }}
     <input type="submit" value="Go">
   </form>
+
+
 </div>
 {% endblock %}
